Fix HypixelLootChestGame being implemented by wrong class
