+++ src/main/java/io/github/hypixel_api_wrapper/wrapper/games/bedwars/HypixelBedWars.java
@@ -1,9 +1,8 @@
 package io.github.hypixel_api_wrapper.wrapper.games.bedwars;
 
-import io.github.hypixel_api_wrapper.wrapper.HypixelLootChestGame;
 import org.json.JSONObject;
 
-public class HypixelBedWars implements HypixelLootChestGame {
+public class HypixelBedWars {
 
     private final String JSONKey;
     private final JSONObject stats;
@@ -94,35 +93,6 @@ public class HypixelBedWars implements HypixelLootChestGame {
     }
 
     public int getCurrentWinstreak() {
-        throw new UnsupportedOperationException();
-    }
-
-    public int getChristmasChestsAmount() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public int getOpenedChestsAmount() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public int getOpenedCommonChestsAmount() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public int getOpenedRareChestsAmount() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public int getOpenedEpicChestsAmount() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public int getOpenedLegendaryChestsAmount() {
         throw new UnsupportedOperationException();
     }
 }
